Move Ember.Checkbox into ember-views package.
