+++ packages/ember-views/lib/main.js
@@ -31,6 +31,7 @@ import Component from "ember-views/views/component";
 import EventDispatcher from "ember-views/system/event_dispatcher";
 import ViewTargetActionSupport from "ember-views/mixins/view_target_action_support";
 import ComponentLookup from "ember-views/component_lookup";
+import Checkbox from "ember-views/views/checkbox";
 // END IMPORTS
 
 /**
@@ -55,6 +56,7 @@ Ember.CoreView = CoreView;
 Ember.View = View;
 Ember.View.states = states;
 Ember.View.cloneStates = cloneStates;
+Ember.Checkbox = Checkbox;
 
 Ember.ComponentLookup = ComponentLookup;
 Ember.ContainerView = ContainerView;
