+++ packages/ember-views/lib/views/checkbox.js
